working on deleteVetByVetId that deletes vet with entities and tests

+++ vet-service/src/main/java/com/petclinic/vet/dataaccesslayer/PhotoRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 public interface PhotoRepository extends ReactiveCrudRepository<Photo, Integer> {
     Mono<Photo> findByVetId(String vetId);
 
-    Mono<Long> deleteByVetId(String vetId);
+    Mono<Integer> deleteByVetId(String vetId);
 }
